2013-02-07 modulo insumos 1.00 (nuevos roles)

+++ ProyectoIntesa/src/edu/client/P_FormularioCliente.java
@@ -1,6 +1,7 @@
 package edu.client;
 
 import java.util.ArrayList;
+import java.util.List;
 
 import com.google.gwt.core.client.GWT;
 import com.google.gwt.event.dom.client.ClickEvent;
@@ -8,15 +9,21 @@ import com.google.gwt.event.dom.client.ClickHandler;
 import com.google.gwt.user.client.Window;
 import com.google.gwt.user.client.rpc.AsyncCallback;
 import com.google.gwt.user.client.ui.Button;
+import com.google.gwt.user.client.ui.ClickListener;
 import com.google.gwt.user.client.ui.Composite;
 import com.google.gwt.user.client.ui.FlexTable;
 import com.google.gwt.user.client.ui.HasHorizontalAlignment;
 import com.google.gwt.user.client.ui.Label;
+import com.google.gwt.user.client.ui.MultiWordSuggestOracle;
 import com.google.gwt.user.client.ui.ScrollPanel;
+import com.google.gwt.user.client.ui.SuggestBox;
 import com.google.gwt.user.client.ui.TabPanel;
 import com.google.gwt.user.client.ui.TextArea;
 import com.google.gwt.user.client.ui.TextBox;
+import com.google.gwt.user.client.ui.Widget;
 
+import edu.client.ComprasService.ComprasService;
+import edu.client.ComprasService.ComprasServiceAsync;
 import edu.client.VentasService.VentasService;
 import edu.client.VentasService.VentasServiceAsync;
 import edu.shared.DTO.ClienteDTO;
@@ -84,16 +91,21 @@ public class P_FormularioCliente extends Composite {
 	private TextBox pisoTb;
 	private TextBox oficinaTb;
 	private TextBox cpaTb;
-	private TextBox localidadTb;
+	private SuggestBox localidadSb;
 	private TextBox codigoPostalTb;
-	private TextBox provinciaTb;
-	private TextBox paisTb;
+	private SuggestBox provinciaSb;
+	private SuggestBox paisSb;
 	private TextBox telefonoTb;
 	private TextBox webTb;
 	private TextBox faxTb;
 	private TextBox emailTb;
 	private TextArea observacionTb;
 
+	private MultiWordSuggestOracle listaLocalidades;
+	private MultiWordSuggestOracle listaProvincias;
+	private MultiWordSuggestOracle listaPaises;
+	
+	
 	// botones
 	private Button btnAgregar;
 	private Button btnCancelar;
@@ -166,10 +178,7 @@ public class P_FormularioCliente extends Composite {
 		pisoTb.setStyleName("textoCorto");
 		oficinaTb = new TextBox();
 		cpaTb = new TextBox();
-		localidadTb = new TextBox();
 		codigoPostalTb = new TextBox();
-		provinciaTb = new TextBox();
-		paisTb = new TextBox();
 		telefonoTb = new TextBox();
 		webTb = new TextBox();
 		faxTb = new TextBox();
@@ -178,6 +187,72 @@ public class P_FormularioCliente extends Composite {
 		observacionTb.setDirectionEstimator(false);
 		observacionTb.setWidth("100%");
 
+		listaLocalidades = new MultiWordSuggestOracle();
+		listaProvincias = new MultiWordSuggestOracle();
+		listaPaises = new MultiWordSuggestOracle();
+		
+		ComprasServiceAsync comprasService = GWT.create(ComprasService.class);
+		comprasService.getNombresPaises(new AsyncCallback<List<String>>() {
+			@Override
+			public void onSuccess(List<String> result) {
+				cargarSugerenciaPaises(result);
+			}
+
+			@Override
+			public void onFailure(Throwable caught) {
+				Window.alert("ERROR EN EL SERVICIO");
+			}
+		});
+		
+			
+		paisSb = new SuggestBox(listaPaises);
+		
+		provinciaSb = new SuggestBox(listaProvincias);
+		provinciaSb.addClickListener(new ClickListener() {
+						
+			@Override
+			public void onClick(Widget sender) {
+				
+				final String pais = paisSb.getText();
+				
+				ComprasServiceAsync comprasService = GWT.create(ComprasService.class);
+				comprasService.getNombresProvincias(pais, new AsyncCallback<List<String>>() {
+					@Override
+					public void onSuccess(List<String> result) {
+						cargarSugerenciaProvincias(result);
+					}
+
+					@Override
+					public void onFailure(Throwable caught) {
+						Window.alert("ERROR EN EL SERVICIO");
+					}
+				});
+			}
+		});
+
+		localidadSb = new SuggestBox(listaLocalidades);
+		localidadSb.addClickListener(new ClickListener() {
+			
+			@Override
+			public void onClick(Widget sender) {
+				
+				final String prov = provinciaSb.getText();
+				
+				ComprasServiceAsync comprasService = GWT.create(ComprasService.class);
+				comprasService.getNombresLocalidades(prov, new AsyncCallback<List<String>>() {
+					@Override
+					public void onSuccess(List<String> result) {
+						cargarSugerenciaLocalidades(result);
+					}
+
+					@Override
+					public void onFailure(Throwable caught) {
+						Window.alert("ERROR EN EL SERVICIO");
+					}
+				});
+			}
+		});
+		
 		btnNuevoContacto = new Button(constante.agregarContacto());
 		btnNuevoContacto.addClickHandler(new ClickHandler() {
 			public void onClick(ClickEvent event) {
@@ -246,11 +321,11 @@ public class P_FormularioCliente extends Composite {
 		formularioCliente.getFlexCellFormatter().setColSpan(3, 0, 10);
 
 		formularioCliente.setWidget(4, 0, pais);
-		formularioCliente.setWidget(4, 1, paisTb);
+		formularioCliente.setWidget(4, 1, paisSb);
 		formularioCliente.setWidget(4, 2, provincia);
-		formularioCliente.setWidget(4, 3, provinciaTb);
+		formularioCliente.setWidget(4, 3, provinciaSb);
 		formularioCliente.setWidget(4, 4, localidad);
-		formularioCliente.setWidget(4, 5, localidadTb);
+		formularioCliente.setWidget(4, 5, localidadSb);
 		formularioCliente.setWidget(4, 6, codigoPostal);
 		formularioCliente.setWidget(4, 7, codigoPostalTb);
 		formularioCliente.setWidget(4, 8, sugerenciaCPA);
@@ -326,6 +401,24 @@ public class P_FormularioCliente extends Composite {
 		this.cliente = clienteSeleccionado;
 		this.titulo = titulo;
 		
+		listaLocalidades = new MultiWordSuggestOracle();
+		listaProvincias = new MultiWordSuggestOracle();
+		listaPaises = new MultiWordSuggestOracle();
+		
+		ComprasServiceAsync comprasService = GWT.create(ComprasService.class);
+		comprasService.getNombresPaises(new AsyncCallback<List<String>>() {
+			@Override
+			public void onSuccess(List<String> result) {
+				cargarSugerenciaPaises(result);
+			}
+
+			@Override
+			public void onFailure(Throwable caught) {
+				Window.alert("ERROR EN EL SERVICIO");
+			}
+		});
+		
+		
 		elementos = new ArrayList<String>();
 		datosEmpresa = new Label(constante.datosEmpresa());
 		datosEmpresa.setStyleName("labelTitulo");
@@ -397,14 +490,14 @@ public class P_FormularioCliente extends Composite {
 		oficinaTb.setText(this.cliente.getDireccion().getOficina());
 		cpaTb = new TextBox();
 		cpaTb.setText(this.cliente.getDireccion().getCpa());
-		localidadTb = new TextBox();
-		localidadTb.setText(this.cliente.getDireccion().getLocalidad());
+		localidadSb = new SuggestBox(listaLocalidades);
+		localidadSb.setText(this.cliente.getDireccion().getLocalidad());
 		codigoPostalTb = new TextBox();
 		codigoPostalTb.setText(this.cliente.getDireccion().getCodigoLocalidad());
-		provinciaTb = new TextBox();
-		provinciaTb.setText(this.cliente.getDireccion().getProvincia());
-		paisTb = new TextBox();
-		paisTb.setText(this.cliente.getDireccion().getPais());
+		provinciaSb = new SuggestBox(listaProvincias);
+		provinciaSb.setText(this.cliente.getDireccion().getProvincia());
+		paisSb = new SuggestBox(listaPaises);
+		paisSb.setText(this.cliente.getDireccion().getPais());
 		telefonoTb = new TextBox();
 		telefonoTb.setText(this.cliente.getTelefono());
 		webTb = new TextBox();
@@ -418,6 +511,50 @@ public class P_FormularioCliente extends Composite {
 		observacionTb.setWidth("100%");
 		observacionTb.setText(this.cliente.getObservaciones());
 
+		provinciaSb.addClickListener(new ClickListener() {
+			
+			@Override
+			public void onClick(Widget sender) {
+				
+				final String pais = paisSb.getText();
+				
+				ComprasServiceAsync comprasService = GWT.create(ComprasService.class);
+				comprasService.getNombresProvincias(pais, new AsyncCallback<List<String>>() {
+					@Override
+					public void onSuccess(List<String> result) {
+						cargarSugerenciaProvincias(result);
+					}
+
+					@Override
+					public void onFailure(Throwable caught) {
+						Window.alert("ERROR EN EL SERVICIO");
+					}
+				});
+			}
+		});
+
+		localidadSb.addClickListener(new ClickListener() {
+			
+			@Override
+			public void onClick(Widget sender) {
+				
+				final String prov = provinciaSb.getText();
+				
+				ComprasServiceAsync comprasService = GWT.create(ComprasService.class);
+				comprasService.getNombresLocalidades(prov, new AsyncCallback<List<String>>() {
+					@Override
+					public void onSuccess(List<String> result) {
+						cargarSugerenciaLocalidades(result);
+					}
+
+					@Override
+					public void onFailure(Throwable caught) {
+						Window.alert("ERROR EN EL SERVICIO");
+					}
+				});
+			}
+		});
+		
 		btnNuevoContacto = new Button(constante.agregarContacto());
 		btnNuevoContacto.addClickHandler(new ClickHandler() {
 			public void onClick(ClickEvent event) {
@@ -487,11 +624,11 @@ public class P_FormularioCliente extends Composite {
 		formularioCliente.getFlexCellFormatter().setColSpan(3, 0, 10);
 
 		formularioCliente.setWidget(4, 0, pais);
-		formularioCliente.setWidget(4, 1, paisTb);
+		formularioCliente.setWidget(4, 1, paisSb);
 		formularioCliente.setWidget(4, 2, provincia);
-		formularioCliente.setWidget(4, 3, provinciaTb);
+		formularioCliente.setWidget(4, 3, provinciaSb);
 		formularioCliente.setWidget(4, 4, localidad);
-		formularioCliente.setWidget(4, 5, localidadTb);
+		formularioCliente.setWidget(4, 5, localidadSb);
 		formularioCliente.setWidget(4, 6, codigoPostal);
 		formularioCliente.setWidget(4, 7, codigoPostalTb);
 		formularioCliente.setWidget(4, 8, sugerenciaCPA);
@@ -561,7 +698,9 @@ public class P_FormularioCliente extends Composite {
 			eliminar.setStyleName("labelBorrar");
 			eliminar.addClickHandler(new ClickHandler() {
 				public void onClick(ClickEvent event) {
-					eliminarContacto(nombreContacto);
+					boolean confirm = Window.confirm("Está seguro de que desea eliminar el contacto?");
+					if(confirm == true)
+						eliminarContacto(nombreContacto);
 				}
 			});
 
@@ -600,17 +739,17 @@ public class P_FormularioCliente extends Composite {
 		boolean vNombreEmpresa = validar.textBoxVacio(this.nombreEmpresaTb.getText());
 		boolean vCuit = validar.textBoxVacio(this.nroCuitTb.getText());
 		boolean vResponsable = validar.textBoxVacio(this.responsableTb.getText());
-		boolean vPais = validar.textBoxVacio(this.paisTb.getText());
-		boolean vProv = validar.textBoxVacio(this.provinciaTb.getText());
-		boolean vLocalidad = validar.textBoxVacio(this.localidadTb.getText());
+		boolean vPais = validar.textBoxVacio(this.paisSb.getText());
+		boolean vProv = validar.textBoxVacio(this.provinciaSb.getText());
+		boolean vLocalidad = validar.textBoxVacio(this.localidadSb.getText());
 		boolean vCalle = validar.textBoxVacio(this.calleTb.getText());
 		boolean vAltura = validar.textBoxVacio(this.alturaTb.getText());
 		
 		if(!vNombreEmpresa && !vCuit && !vResponsable && !vPais && !vProv && !vLocalidad && !vCalle && !vAltura){
 			DireccionDTO direccion = new DireccionDTO();
-			direccion.setPais(this.paisTb.getText());
-			direccion.setProvincia(this.provinciaTb.getText());
-			direccion.setLocalidad(this.localidadTb.getText());
+			direccion.setPais(this.paisSb.getText());
+			direccion.setProvincia(this.provinciaSb.getText());
+			direccion.setLocalidad(this.localidadSb.getText());
 			direccion.setCodigoLocalidad(this.codigoPostalTb.getText());
 			direccion.setCalle(this.calleTb.getText());
 			direccion.setAltura(this.alturaTb.getText());
@@ -683,17 +822,17 @@ public class P_FormularioCliente extends Composite {
 		
 		boolean vCuit = validar.textBoxVacio(this.nroCuitTb.getText());
 		boolean vResponsable = validar.textBoxVacio(this.responsableTb.getText());
-		boolean vPais = validar.textBoxVacio(this.paisTb.getText());
-		boolean vProv = validar.textBoxVacio(this.provinciaTb.getText());
-		boolean vLocalidad = validar.textBoxVacio(this.localidadTb.getText());
+		boolean vPais = validar.textBoxVacio(this.paisSb.getText());
+		boolean vProv = validar.textBoxVacio(this.provinciaSb.getText());
+		boolean vLocalidad = validar.textBoxVacio(this.localidadSb.getText());
 		boolean vCalle = validar.textBoxVacio(this.calleTb.getText());
 		boolean vAltura = validar.textBoxVacio(this.alturaTb.getText());
 		
 		if(!vCuit && !vResponsable && !vPais && !vProv && !vLocalidad && !vCalle && !vAltura){
 			DireccionDTO direccion = new DireccionDTO();
-			direccion.setPais(this.paisTb.getText());
-			direccion.setProvincia(this.provinciaTb.getText());
-			direccion.setLocalidad(this.localidadTb.getText());
+			direccion.setPais(this.paisSb.getText());
+			direccion.setProvincia(this.provinciaSb.getText());
+			direccion.setLocalidad(this.localidadSb.getText());
 			direccion.setCodigoLocalidad(this.codigoPostalTb.getText());
 			direccion.setCalle(this.calleTb.getText());
 			direccion.setAltura(this.alturaTb.getText());
@@ -771,7 +910,9 @@ public class P_FormularioCliente extends Composite {
 		eliminar.setStyleName("labelBorrar");
 		eliminar.addClickHandler(new ClickHandler() {
 			public void onClick(ClickEvent event) {
-				eliminar(nombreContacto);
+				boolean confirm = Window.confirm("Está seguro de que desea eliminar el contacto?");
+				if(confirm == true)
+					eliminar(nombreContacto);
 			}
 		});
 
@@ -839,6 +980,29 @@ public class P_FormularioCliente extends Composite {
 		}
 
 		return elemento;
+	}
+	
+	protected void cargarSugerenciaPaises(List<String> result) {
+		for (String sugerencia : result) {
+			listaPaises.add(sugerencia);
+		}
+		
+	}
+	
+	protected void cargarSugerenciaProvincias(List<String> result) {
+		listaProvincias.clear();
+		for (String sugerencia : result) {
+			listaProvincias.add(sugerencia);
+		}
+		
+	}
+	
+	protected void cargarSugerenciaLocalidades(List<String> result) {
+		listaLocalidades.clear();
+		for (String sugerencia : result) {
+			listaLocalidades.add(sugerencia);
+		}
+		
 	}
 
 }
